Update timeline.php
Previously: All circles had a fixed length regardless of the year gaps.

Fix: Found a reasonable range of values that were used to represent how far the circles were and then normalized them in a range of 0 - 1 so that they would all fit on the timeline according to their given ranges. The timeline’s full length would be represented as a 1.

Also, centered the images again to match this.