Test Header which includes screen size bug fixes
The previous header was only suited for larger devices such as laptops. Bugs included improper search bar formatting, links and search bar were not responsive, links disappeared based on screen size and the logo was not uploaded. 

This is a test file as more changes will need to be made for new login and log out pages and header will need to made into a PHP file