Museum Data and Research for Timeline
All research and data were gathered by Sowmya Movva 

The document is 26 pages excluding the iCloud folder of exhibits with detailed dates and text which did not require further research and cross-checking with the museum database. Data was conducted over the course of one month with further refinements taking place mid-March. 

Additionally, this research was used to create graphical elements in Figma where each image is used as the front of an event card in the timeline and corresponding information representing the rear part of the photocard. Figma has additional information that was obtained from iCloud.